Refactor CRM integration
