+++ Http/Controllers/AmeiseController.php
@@ -8,18 +8,24 @@ use Carbon\Carbon;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
 use Illuminate\Routing\Controller;
-use Modules\AmeiseModule\Services\CrmService;
+use Modules\AmeiseModule\Services\TokenService;
+use Modules\AmeiseModule\Services\CrmApiClient;
+use Modules\AmeiseModule\Services\ConversationArchiver;
 use Modules\AmeiseModule\Entities\CrmArchive;
 use Modules\AmeiseModule\Entities\CrmArchiveThread;
 
 class AmeiseController extends Controller
 {
-    protected $crmService;
+    protected $tokenService;
+    protected $apiClient;
+    protected $archiver;
+
     public function __construct()
     {
-
         $this->middleware(function ($request, $next) {
-            $this->crmService = $this->crmService ?? new CrmService('', auth()->user()->id);
+            $this->tokenService = $this->tokenService ?? new TokenService('', auth()->user()->id);
+            $this->apiClient = new CrmApiClient($this->tokenService);
+            $this->archiver = new ConversationArchiver($this->apiClient);
             return $next($request);
         });
 
@@ -34,18 +40,18 @@ class AmeiseController extends Controller
             case 'crm_users_search':
                 $results = [];
                 if(!empty($inputs['new_conversation'])) {
-                    $results = $this->crmService->getFSUsers($inputs);
+                    $results = $this->getFSUsers($inputs);
                 }
-                return $this->crmService->getCrmUsers($inputs, $results);
+                return $this->getCrmUsers($inputs, $results);
                 break;
 
             case 'get_contract':
-                $response = $this->crmService->getContracts($request->input('client_id'));
+                $response = $this->apiClient->getContracts($request->input('client_id'));
                 if (isset($response['error']) && isset($response['url'])) {
                     return response()->json(['error' => 'Redirect', 'url' => $response['url']]);
                 }
-                $divisionResponse = $this->crmService->getContactEndPoints('sparten');
-                $statusResponse = $this->crmService->getContactEndPoints('vertragsstatus');
+                $divisionResponse = $this->apiClient->getContactEndPoints('sparten');
+                $statusResponse = $this->apiClient->getContactEndPoints('vertragsstatus');
                 $groupedData = collect($response)->groupBy('Status')->map(function ($group) use ($divisionResponse, $statusResponse) {
                     return $group->map(function ($items) use ($divisionResponse, $statusResponse) {
                         $divisionKey = array_search($items['Sparte'], array_column($divisionResponse, 'Value'));
@@ -86,9 +92,9 @@ class AmeiseController extends Controller
                             $crm_user_id = $inputs['customer_id'];
                             $contracts = json_decode($inputs['contracts'], true);
                             $divisions = json_decode($inputs['divisions_data'], true);
-                            $conversation_data = $this->crmService->createConversationData($conversation, $crm_user_id, $contracts, $divisions, $thread);
-                            if($this->crmService->archiveConversation($conversation_data)) {                        
-                                $this->crmService->archiveConversationWithAttachments($thread, $conversation_data);
+                            $conversation_data = $this->archiver->createConversationData($conversation, $crm_user_id, $contracts, $divisions, $thread);
+                            if($this->apiClient->archiveConversation($conversation_data)) {
+                                $this->archiver->archiveConversationWithAttachments($thread, $conversation_data);
                                 CrmArchiveThread::create(['crm_archive_id' => $crm_archive->id,'thread_id' => $thread->id,'conversation_id'=> $conversation->id ]);
                             }
                         }
@@ -114,5 +120,62 @@ class AmeiseController extends Controller
             'archives' => $archives,
         ])->render();
 
+    }
+
+    private function getCrmUsers($inputs, $result = [])
+    {
+        $response = $this->apiClient->fetchUserByIdOrName($inputs['search']);
+        if (isset($response['error']) && isset($response['url'])) {
+            return response()->json(['error' => 'Redirect', 'url' => $response['url']]);
+        }
+        $crmUsers = [];
+        foreach($response as $data) {
+            $emails = $phone =  [];
+            $contactDetails = $this->apiClient->fetchUserDetail($data['Id'], 'kontaktdaten');
+            foreach ($contactDetails as $item) {
+                if ($item["Typ"] === "email") {
+                    $emails[] = $item["Value"];
+                } elseif($item['Typ'] == 'telefon') {
+                    $phone [] = $item['Value'];
+                }
+            }
+            $crmUsers[] = [
+                'id' => $data['Id'],
+                'text' => $data['Text'],
+                'id_name' => $data['Person']['Vorname'] . " " . $data['Person']['Nachname'] . "(" . $data['Id'] . ")",
+                'first_name' => $data['Person']['Vorname'],
+                'last_name'  => $data['Person']['Nachname'],
+                'address'    => $data['Hauptwohnsitz']['Strasse'],
+                'zip'        => $data['Hauptwohnsitz']['Postleitzahl'],
+                'city'       => $data['Hauptwohnsitz']['Ort'],
+                'country'    => $data['Hauptwohnsitz']['Land'],
+                'emails'     => $emails,
+                'phones'     => $phone,
+            ];
+        }
+        $result['crmUsers'] = $crmUsers;
+        return response()->json($result);
+    }
+
+    private function getFSUsers($inputs)
+    {
+        $response = [];
+        $q = $inputs['search'];
+        $customers_query = \App\Customer::select(['customers.id', 'first_name', 'last_name', 'emails.email'])->join('emails', 'customers.id', '=', 'emails.customer_id');
+        $customers_query->where('emails.email', 'like', '%'.$q.'%');
+        $customers_query->orWhere('first_name', 'like', '%'.$q.'%')
+            ->orWhere('last_name', 'like', '%'.$q.'%');
+        $customers = $customers_query->paginate(20);
+        foreach ($customers as $customer) {
+            $id = '';
+            $text = $customer->getNameAndEmail();
+            $id = $customer->email;
+
+            $response['fsUsers'][] = [
+                'id'   => $id,
+                'text' => $text,
+            ];
+        }
+        return $response;
     }
 }
